structural changes:separated by store router
